TIP-885: move Format controller
