+++ src/Pim/Bundle/LocalizationBundle/DependencyInjection/PimLocalizationExtension.php
@@ -22,8 +22,5 @@ class PimLocalizationExtension extends Extension
         $configuration = new Configuration();
         $config = $this->processConfiguration($configuration, $configs);
         $container->prependExtensionConfig('pim_localization', $config);
-
-        $loader = new YamlFileLoader($container, new FileLocator(__DIR__.'/../Resources/config'));
-        $loader->load('controllers.yml');
     }
 }
