--- src/Pim/Bundle/LocalizationBundle/Resources/config/controllers.yml
@@ -1,14 +1,0 @@
-parameters:
-    pim_localization.controller.format.class: Pim\Bundle\LocalizationBundle\Controller\FormatController
-
-services:
-    pim_localization.controller.format:
-        class: '%pim_localization.controller.format.class%'
-        arguments:
-            - '@pim_catalog.localization.factory.date'
-            - '@pim_catalog.localization.factory.datetime'
-            - '@pim_enrich.resolver.locale'
-            - '@pim_user.context.user'
-            -
-                decimal_separators: '%pim_catalog.localization.decimal_separators%'
-                date_formats: '%pim_catalog.localization.date_formats%'
